Move `Bounded3d` for `Triangle3d`
